+++ components/ocs_http/target_esp32/server.h
@@ -23,10 +23,10 @@ class Server : public IServer,
 public:
     struct Params {
         //! TCP port to accept incoming connections.
-        unsigned server_port { 80 };
+        uint8_t server_port { 80 };
 
         //! Maximum allowed URI handlers.
-        unsigned max_uri_handlers { 32 };
+        uint8_t max_uri_handlers { 32 };
     };
 
     //! Initialize.
